+++ src/main/resources/plugin.yml
@@ -1,0 +1,32 @@
+name: Economy
+main: me.scruffyboy13.Economy.Economy
+version: 1.7.3
+author: scruffyboy13
+depend: [Vault]
+
+commands:
+  money:
+    aliases: [economy, eco]
+  balance:
+    aliases: [bal]
+  balancetop:
+    aliases: [baltop]
+  pay:
+
+permissions:
+  economy.command.balance:
+    default: true
+  economy.command.balancetop:
+    default: true
+  economy.command.help:
+    default: true
+  economy.command.pay:
+    default: true
+  economy.command.give:
+    default: op
+  economy.command.take:
+    default: op
+  economy.command.set:
+    default: op
+  economy.command.reload:
+    default: op
